Important update of the MiniZincModel class. Now the model takes care of the naming, the serialization and creation of the mzn file. Now I assume that the input of mzn2fzn is always a file, and the output_base is always the name of the input_file. If one wants to redirect the output into another location, then it should provide an output_base to the minizinc function or to the MiniZincModel instance, which creates a new mzn file in the specified output_base and thus the fzn and ozn. Now the mzn2fzn function returns only the fzn and ozn. no_ozn option removed. serialization is now not enabled by default, unless one provides a model file to a minizinc model and modifications are made on it.

+++ pymzn/mzn/minizinc.py
@@ -1,20 +1,18 @@
-import os
-import logging
-import itertools
 import contextlib
+import logging
+import os
 from subprocess import CalledProcessError
 
 import pymzn.config as config
-from pymzn.dzn import parse_dzn, dzn
 from pymzn.bin import cmd, run
-from pymzn.mzn.solvers import gecode
+from pymzn.dzn import parse_dzn, dzn
 from pymzn.mzn.model import MiniZincModel
+from pymzn.mzn.solvers import gecode
+
+# TODO: check all the docstrings
 
 
-_minizinc_instance_counter = itertools.count()
-
-
-def minizinc(mzn, *dzn_files, data=None,
+def minizinc(mzn, *dzn_files, data=None, serialize=False, raw_output=False,
              keep=False, output_base=None, output_vars=None,
              mzn_globals_dir='gecode', fzn_fn=gecode, **fzn_args):
     """
@@ -67,40 +65,21 @@ def minizinc(mzn, *dzn_files, data=None,
 
     if isinstance(mzn, MiniZincModel):
         mzn_model = mzn
-    elif isinstance(mzn, str):
-        mzn_model = MiniZincModel(mzn)
     else:
-        raise TypeError('The input _model is invalid.')
+        mzn_model = MiniZincModel(mzn, output_base=output_base,
+                                  serialize=serialize)
+    if not raw_output:
+        mzn_model.dzn_output_stmt(output_vars)
+    mzn_file = mzn_model.compile()
 
-    # TODO: separate the handling of model related reads and writes from the
-    # minizinc function
-    mzn_model.dzn_output_stmt(output_vars)
-    mzn = mzn_model.compile()
-
-    # Ensures isolation of instances and thread safety
-    global _minizinc_instance_counter
-    instance_number = next(_minizinc_instance_counter)
-
-    output_base = output_base or mzn_model.mzn_out_file[:-4]
-    output_base = '{}_{}'.format(output_base, instance_number)
-
-    mzn_file = output_base + '.mzn'
     try:
-        # Execute mzn2fzn
-        mzn_file, fzn_file, ozn_file = mzn2fzn(mzn, data=data,
-                                               dzn_files=dzn_files,
-                                               output_base=output_base,
-                                               mzn_globals_dir=mzn_globals_dir)
+        fzn_file, ozn_file = mzn2fzn(mzn_file, *dzn_files, data=data,
+                                     mzn_globals_dir=mzn_globals_dir)
         try:
-            # Execute fzn_fn
-            fzn_args = fzn_args or {}
             solns = fzn_fn(fzn_file, **fzn_args)
-
-            # Execute solns2out
             out = solns2out(solns, ozn_file)
-
-            # Parse the solutions
-            out = map(parse_dzn, out)
+            # TODO: check if stream-ability possible now, in case remove list
+            return list(map(parse_dzn, out))
         finally:
             if not keep:
                 with contextlib.suppress(FileNotFoundError):
@@ -113,18 +92,18 @@ def minizinc(mzn, *dzn_files, data=None,
     finally:
         if not keep:
             with contextlib.suppress(FileNotFoundError):
-                os.remove(mzn_file)
-                log.debug('Deleting file: %s', mzn_file)
-    return out
+                if mzn_file:
+                    os.remove(mzn_file)
+                    log.debug('Deleting file: %s', mzn_file)
 
 
-def mzn2fzn(mzn, dzn_files=None, *, data=None, output_base=None, no_ozn=False,
-            mzn_globals_dir='gecode'):
+def mzn2fzn(mzn_file, *dzn_files, data=None, mzn_globals_dir='gecode'):
     """
     Flatten a MiniZinc _model into a FlatZinc one. It executes the mzn2fzn
     utility from libminizinc to produce a fzn and ozn files from a mzn one.
 
-    :param str mzn: The path to a mzn file containing the MiniZinc _model or
+    :param str mzn_file: The path to a mzn file containing the MiniZinc
+    _model or
                     the content of the _model.
     :param [str] dzn_files: A list of paths to dzn files to attach to the
                             mzn2fzn execution; by default no data file is
@@ -138,52 +117,20 @@ def mzn2fzn(mzn, dzn_files=None, *, data=None, output_base=None, no_ozn=False,
                             is used to name the file where the mzn _model
                             will be written. In that case, if output_base is
                             None then a default name ('mznout') is used.
-    :param bool no_ozn: Whether to create the ozn file or not. Default is
-                        False (create). If no ozn is created, it is still
-                        possible to use solns2out to parse the solution
-                        stream output of the solver. Notice though that
-                        MiniZinc optimizes the _model also according to its
-                        output so it is recommended to use it (if the
-                        minizinc function is used, it is recommended to use
-                        a _model in which replace_output_stmt=True, default
-                        behaviour)
     :param str mzn_globals_dir: The name of the directory to search for
                                 globals included files in the standard
                                 library; by default the 'gecode' global
                                 library is used, since Pymzn assumes Gecode
                                 as default solver
     :return: The paths to the mzn, fzn and ozn files created by the function
-    :rtype: (str, str, str)
+    :rtype: (str, str)
     """
     log = logging.getLogger(__name__)
 
-    if not isinstance(mzn, str):
-        raise ValueError('The input _model must be a string.')
-
-    if mzn.endswith('.mzn'):
-        mzn_file = mzn
-        log.debug('Mzn file provided: %s', mzn_file)
-    else:
-        if output_base:
-            mzn_file = ''.join([output_base, '.mzn'])
-            output_base = None
-        else:
-            mzn_file = 'mznout.mzn'
-
-        log.debug('Writing provided content to: %s', mzn_file)
-        with open(mzn_file, 'w') as f:
-            f.write(mzn)
-
     args = []
-
-    if output_base:
-        args.append(('--output-base', output_base))
 
     if mzn_globals_dir:
         args.append(('-G', mzn_globals_dir))
-
-    if no_ozn:
-        args.append('--no-output-ozn')
 
     if data is not None:
         data = '"{}"'.format(' '.join(dzn(data)))
@@ -200,17 +147,17 @@ def mzn2fzn(mzn, dzn_files=None, *, data=None, output_base=None, no_ozn=False,
         log.exception(err.stderr)
         raise RuntimeError(err.stderr) from err
 
-    base = output_base or mzn_file[:-4]
+    base = os.path.splitext(mzn_file)[0]
 
     fzn_file = '.'.join([base, 'fzn'])
     if not os.path.isfile(fzn_file):
         fzn_file = None
 
     ozn_file = '.'.join([base, 'ozn'])
-    if no_ozn or not os.path.isfile(ozn_file):
+    if not os.path.isfile(ozn_file):
         ozn_file = None
 
-    return mzn_file, fzn_file, ozn_file
+    return fzn_file, ozn_file
 
 
 def solns2out(solns_input, ozn_file=None):
